requirements.txt
requests==2.32.3
urllib3==2.4.0
python-dotenv==1.1.0